+++ lib/pork.rb
@@ -60,9 +60,9 @@ module Pork
     @auto = auto
     @autorun ||= at_exit do
       next unless @auto
-      Random.srand(ENV['PORK_SEED'].to_i) if ENV['PORK_SEED']
-      execute_mode(ENV['PORK_MODE'])      if ENV['PORK_MODE']
-      require "pork/mode/#{execute_mode}" if execute_mode.to_s != 'execute'
+      Random.srand(ENV['PORK_SEED'].to_i)   if ENV['PORK_SEED']
+      execute_mode(ENV['PORK_MODE'].to_sym) if ENV['PORK_MODE']
+      require "pork/mode/#{execute_mode}"   if execute_mode != :execute
       seed
       trap
       run